feat: Add copy-tab-info functionality to background and popup
The code changes add the functionality to copy the title and URL of the active tab in the browser. The background script listens for the "copy-tab-info" command and retrieves the tab information, while the popup script displays the tab information and allows the user to copy it to the clipboard. Additionally, the popup script retrieves and displays the current shortcut key for the "copy-tab-info" command.

Note: The commit message follows the conventional format of "feat: <description>".