feat: Add functionality to copy tab information without parameters
This commit adds the ability to copy the title and URL of the active tab in the browser without including any query parameters. It introduces a new command "copy-tab-info-no-params" that retrieves the tab information and displays it in the popup. The shortcut key for this command is also added to the options page.

Note: The commit message follows the conventional format of "feat: <description>".